+++ tornadose/handlers.py
@@ -1,7 +1,7 @@
 """Custom request handlers for pushing data to connected clients."""
 
 from tornado import gen
-from tornado.web import RequestHandler, HTTPError
+from tornado.web import RequestHandler
 from tornado.websocket import WebSocketHandler, WebSocketClosedError
 from tornado.queues import Queue
 from tornado.iostream import StreamClosedError
@@ -9,8 +9,25 @@ from tornado.log import access_log
 from . import stores
 
 
-class EventSource(RequestHandler):
-    """Base handler for server-sent events a.k.a. EventSource.
+class BaseHandler(RequestHandler):
+    """Base handler for subscribers."""
+    def submit(self, message):
+        """Submit a new message to be published. This method must be
+        implemented by child classes.
+
+        """
+        raise NotImplementedError('submit must be implemented!')
+
+    def publish(self):
+        """Push a message to the subscriber. This method must be
+        implemented by child classes.
+
+        """
+        raise NotImplementedError('publish must be implemented!')
+
+
+class EventSource(BaseHandler):
+    """Handler for server-sent events a.k.a. EventSource.
 
     The EventSource__ interface has a few advantages over websockets:
 
@@ -26,20 +43,16 @@ class EventSource(RequestHandler):
     __ https://github.com/jkbrzt/httpie
 
     """
-    def initialize(self, source, period=None):
-        """The ``source`` parameter is a string that is updated with
-        new data. The :class:`EventSource` instance will continuously
-        check if it is updated and publish to clients when it is.
-
-        If ``period`` is given, publishers will sleep for approximately the
-        given time in order to throttle data speeds.
+    def initialize(self, store, period=None):
+        """If ``period`` is given, publishers will sleep for
+        approximately the given time in order to throttle data
+        speeds.
 
         """
-        assert isinstance(source, (stores.DataStore, stores.StoreContainer))
+        assert isinstance(store, stores.BaseStore)
         assert isinstance(period, (int, float)) or period is None
-        self.source = source
+        store.register(self)
         self.period = period
-        self._last = None
         self.finished = False
         self.set_header('content-type', 'text/event-stream')
         self.set_header('cache-control', 'no-cache')
@@ -52,6 +65,12 @@ class EventSource(RequestHandler):
             self._request_summary(), request_time)
 
     @gen.coroutine
+    def submit(self, message):
+        """Receive incoming data."""
+        yield self.publish(message)
+
+
+    @gen.coroutine
     def publish(self, message):
         """Pushes data to a listener."""
         try:
@@ -62,42 +81,42 @@ class EventSource(RequestHandler):
 
     @gen.coroutine
     def get(self, *args, **kwargs):
-        if type(self.source) is stores.StoreContainer:
-            index = int(args[0])
-            if index >= len(self.source):
-                raise HTTPError(405, 'StoreContainer index out of range')
-            self.source = self.source[index]
-
-        while not self.finished:
-            if self.source.id != self._last and self.source.data is not None:
-                yield self.publish(self.source.data)
-                self._last = self.source.id
-            else:
+        try:
+            while not self.finished:
                 if self.period is not None:
                     yield gen.sleep(self.period)
-        self.finish()
+                else:
+                    yield gen.moment
+        except Exception:
+            pass
+        finally:
+            self.store.deregister(self)
+            self.finish()
 
 
-class WebSocketSubscriber(WebSocketHandler):
+class WebSocketSubscriber(BaseHandler, WebSocketHandler):
     """A Websocket-based subscription handler to be used with
-    :class:`tornadose.stores.Publisher`.
+    :class:`tornadose.stores.QueueStore`.
 
     """
-    def initialize(self, publisher):
-        self.publisher = publisher
+    def initialize(self, store):
+        self.store = store
         self.messages = Queue()
         self.finished = False
 
+    @gen.coroutine
     def open(self):
         """Register with the publisher."""
-        self.publisher.register(self)
-        self.run()
+        self.store.register(self)
+        while not self.finished:
+            message = yield self.messages.get()
+            yield self.publish(message)
 
     def on_close(self):
         self._close()
 
     def _close(self):
-        self.publisher.deregister(self)
+        self.store.deregister(self)
         self.finished = True
 
     @gen.coroutine
@@ -105,15 +124,6 @@ class WebSocketSubscriber(WebSocketHandler):
         yield self.messages.put(message)
 
     @gen.coroutine
-    def run(self):
-        """Coroutine for publishing messages as they become
-        available.
-
-        """
-        while not self.finished:
-            message = yield self.messages.get()
-            self.publish(message)
-
     def publish(self, message):
         """Push a new message to the client. The data will be
         available as a JSON object with the key ``data``.